Introduce Compactor configuration for dealing with consecutive failures  (#5726)
This change propagates exceptions from the
classloading related code. Prior to this change
some of the exceptions raised by the classloading
code would be caught early and a
RuntimeException would be raised instead.

This change also modifies the Compactor to
conditionally delay execution of the next
compaction job when the Compactors has
been failing to complete consecutive prior
compactions. Four new properties control
the behavior of new delay logic.

This change also modifies the API between
the Compactor and Coordinator when
compactions have failed. The exception class
is now relayed to the coordinator, which is
tracking and periodically logging a summary
of the failures.

New metrics related to outcome of the
compaction on the compactor have been added.


Co-authored-by: Keith Turner <kturner@apache.org>