Created Resource Group configuration
This change adds support for persistent configurations
for Resource Groups. On instance upgrade or initialization
the /resource_groups path is created in ZooKeeper and the
default resource group node is created /resource_groups/default.

AccumuloClient.resourceGroupOperations() returns an instance of
ResourceGroupOperations which has methods for creating and
removing Resource Group configuration nodes and for setting,
modifying, and deleting properties. The Shell config command,
zoo-prop-viewer and zoo-prop-editor have been modified to
accept Resource Group arguments.

The SystemConfiguration class has been modified to merge the
Resource Group configuration for the server process with the
system configuration, giving priority to the Resource Group
configuration value, if defined.

ResourceGroupPropUtil.validateResourceGroupProperty restricts
which properties can be set in the Resource Group configuration.