avoids reassignment of hosted tablet during shutdown (#5751)
When a tablet server is gracefully shutdown it notifies the manager it
is shutting down. The manager would remove the server from the set of
live tservers which caused the manger to treat all tablets on that
tserver as assigned to dead tservers.

Saw this when debugging a failure with GracefulShutdownIT. Saw the
manager initiate log recovery and reassignment for the root tablet
before it unloaded from the tserver that was shutting down. This caused
errors with the metadata tablet consistency check during unload of the
root tablet because the manager had modified the tablet metadata before
unload.

This fix modifies the tablet server to use the existing fate operation
for shutting down a tserver.