adds time outs to coordinator tserver RPCs
The coordinator was making RPCs to tservers w/ no timeout set on the TCP
connection. This could cause the coordinator to get stuck indefinitely on
problematic tservers causing external compactions to stop.  Added a
timeout to the connection.  Also modified the thread name to include the
tserver when making a RPC.  Also made the coordinator more aggressive
when it does see a tserver RPC error, made it clear all summary info
related to that tserver instead of a subset.  This more aggresive
clearing is meant to minimize threads going to a problem tserver.