Adds per table and compaction erasure code support (#5743)
Adds Erasure Codes support in Accumulo.  The changes to RFileOperations
allow EC to be treated similarly to replication.  Two new table
properties are introduced.  One to enable EC on a per table basis and
one to set the specific ec policy for the table.  CompressionConfigurer
was extended with ErasureCodeConfigurer as an example for more
specialized ec handling.  It will enforce constraints on ec.  File size
limits can be imposed to encode only larger files.


Co-authored-by: cawaring <cawaring@apache.org>
Co-authored-by: Dave Marion <dlmarion@apache.org>