adds group summary information to reports. (#5779)
* Adds group summary info to print and json reports

Adds the group summaries to the json object and the printed report for
more accurate system status reports.

Also ensures the --showHosts flag works on the json object and printed
report.

* Removes csv output option

Removes the csv output option as the output format is difficult to
maintain and is not structured in a descriptive way.