Considers projected compactions when searching for files to compact (#5675)
When attempting to merge #5588 to main, found a bug with the change.
The bug was that the code did not consider transitive compactions when
searching past currently running compactions.  This bug would not cause
any errors, it was just suboptimal. After these changes no compactions
that would transitively include any smaller files will be planned.

For a tablet that has 6 1M files compacting, 5 7M files and 4 40M files
the code without this fix would find the 4 40M files to compact.
However if we wait for the compaction of the 5 7M files (which is
waiting on the compaction of the 6 1M files), then the 4 40M files would
be compacted with the 41M file that would be produced. These changes
would wait for that making the three compactions run sequentially so they
can include each others output.

A second test added in this change gives examples cases of compactions
that can be found when there are running compaction.