Fix FatePoolsWatcherITBase debugging (#5804)
Was catching exceptions and hiding assertion failures, but was using a
boolean to check that all assertions occurred. No longer catch
exceptions in the test, allowing us to remove the all assertions
occurred check.

This is just a follow on to PR #5790 which introduced these changes for
one test but not all