Fixed issue with exception in server run method not being logged (#5796)
The use of try-with-resources in the server process' main method
causes the server's close method to always be called. When an
exception occurs in the runServer method it may not be logged
because the invocation of the close method consumes enough time
that the lock watcher is fired which ends up halting the JVM. This
change ensures that exceptions from the servers `run` and `close`
methods are logged.

Closes #5794