avoid loss of error thrown from server run()
Follow up to #5796 that attempts to handle the following case.

 1. server.runServer() throws an error
 2. server.close() throws en exception

When the above happens the error will never be logged and its also not
attached to the exception thrown in the finally block by server.close().
Therefore the error is lost and there is no indication it happeneded.
With these changes the exception thrown by server.close() will be
attached to the error thrown by server.runServer().