avoid loss of error thrown from server run() (#5808)
Follow up to #5796 that attempts to handle the following case.

 1. server.runServer() throws an error
 2. server.close() throws en exception or error

When the above happens the error will never be logged and its also not
attached to the exception thrown in the finally block by server.close().
Therefore the error is lost and there is no indication it happened. Changed 
the code to catch throwable to avoid loss of information in this case.


Co-authored-by: Dave Marion <dlmarion@apache.org>