Added compaction service configuration validation to upgrade --start (#5807)
Added compaction service configuration validation to UpgradeUtil.
When `accumulo upgrade --start` is run it will fail if any of the
configured compaction services do not exist for tables other than
the root and metadata table. For the root and metadata table, it will
log a warning that, if not fixed, the compaction service configurations
will be removed from the root and metadata tables on the upgrade so
that they can use the default compaction service configuration.
Modified the Upgrader to implement this change.

Closes #5805