Add TransactionRunner to set when created, not when first run
The TransactionRunner instance was being added to the
runningTxRunners when the TransactionRunner.run method
is called, not when it's created. FateExecutor.initiateShutdown
iterates over runningTxRunners and calls flagStop to
tell the TransactionRunners to exit. Before this fix
it's possible that a TransactionRunner could be created,
but not run by the ThreadPoolExecutor, and therefore
miss the stop signal.