Added resource group parameter to client Thrift connection methods (#5775)
The new resource group parameter has been added in anticipation
of the resource operations api addition in #5749. This change
also adds a debug system property like we have for hosts that
allows the user to configure the client to only connect to
hosts in the resource group for non-management api operations.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>