Add Sunny tag to ComprehensiveTableOperationsIT_SimpleSuite
making it such that there were no ITs that get run as part
of SimpleSharedMacTestSuiteIT with the sunny profile. This
caused the sunny profile test run to fail because the
SimpleSharedMacTestSuiteIT would not find any tests to run
and fail. Adding the sunny tag to one test that runs as
part of the suite fixes the issue.

Closes #5851