Allow fixed properties to be set in resource group and other fixes
Modified ClusterConfigParser to create the resource groups directly
in ZooKeeper vs using the API. When the API is used, then it requires
the Manager process to be up and running. This may not be the case
when starting the instance via accumulo-cluster.

Modified ResourceGroupOperationsImpl to use ResourceGroupPredicate.ANY
when getting properties for the resource group. The properties are
stored in ZK, so any process should be able to retrieve them. Only
when getting the configuration for the resource group do we need to
connect to a process running in that resource group.

The remainder of the changes are to enable fixed properties to be
set in the resource group and perform appropriate warning in the
shell.

Related to #5859