FATE config changes (#5861)
* FATE config changes

changes the FATE config structure

From:
* being a JSON where each key is a comma-separated list of FATE operations and each value is a pool size for those operations
To:
* being a JSON where each key is some user-defined name and each value is a JSON with a single key/value where the key is a comma-separated list of FATE operations and the value is a pool size for those operations
For example:
* {"SOME_OPS": 4, "REMAINING_OPS": 10}
Could become:
* {"pool1": {"SOME_OPS": 4}, "pool2": {"REMAINING_OPS": 10}}

This was done to facilitate some way to maintain metrics for these pools without including the (potentially very long) list of FATE operations, which could be an issue for some monitoring systems. Also shortens some log messages. Instead of the FateExecutor objects only being identifiable by their assigned FATE ops, they are now given the name as defined in the config which can be used to identify them as well.

Previously:
* metrics tags would include all the FATE operations for that pool, but these long metrics could be an issue for some monitoring systems
* the pool name would be "accumulo.pool.manager.fate.[user/meta].[THE OPS]"
* the work finder for the pool would be named "fate.work.finder.[user/meta].[THE OPS]"

Now:
* metrics tags include the name instead of the entire list of operations (tag looks like "accumulo.pool.manager.fate.[user/meta].[THE NAME]")
* the pool name is "accumulo.pool.manager.fate.[user/meta].[THE NAME]"
* the work finder thread for the pool is named "fate.work.finder.[user/meta].[THE NAME]"

closes #5836