Use ServerId to represent a server address
As part of #5775 ThriftTransportPool.getAnyCachedTransport was modified
to perform a ZooKeeper lookup to get the ResourceGroup for a server
because that information is not available on the ThriftTransportKey.
ThriftTransportKey contains the HostAndPort only. The reason for this
change is to make the ResourceGroup available for the ResourceGroupPredicate
that is passed into `ThriftTransportPool.getAnyCachedTransport` so that
the ZooKeeper lookup can be removed.

The server address is currently represented by either a String or a
HostAndPort object in the code. Additionally the addresses are persisted
in file paths in HDFS and in the root and metadata table. I opted to use
the ServerId object as a replacement as it already contains the necessary
information and is part of the public API, but making the change touched
a lot of code.

Closes #5848