Adds tablet extent, files, and full stacktrace to log message (#5863)
* Adds tabletID, files, and full stacktrace to log

Adds the tabletID, the set of files being called, and the full stack
trace to the log message for better troubleshooting.

If an exception is returned for "File does not exist" then compute the
file set difference and log the output.
If there is not a set difference than still log an error.