Fixed a FunctionalTestUtils method
Improves usage of FunctionalTestUtils.getStoredTabletFiles(). Was previously always scanning the METADATA table for stored tablet files of the provided table. This doesn't work if we want the stored tablet files of the METADATA table or the ROOT table, in which case we should be scanning the ROOT table or the root tablet metadata (in ZK), respectively. Used Ample to achieve this.