shortens help description for accumulo upgrade (#5887)
When running `accumulo -help` the description of the upgrade command is
really long compared to other commands.  Shortened the description by
moving information into the options so that the more detailed
information is displayed when running `accumulo upgrade --help`.  Some
of the information in the description was redundant w/ the existing help
for `--prepare` and did not need to be moved.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>