Warn when fixed property set in resource group, other fixes (#5865)
Created new `accumulo init  --add-resource-groups` command to
enable users to define resource groups in ZooKeeper before trying
to start processes in them. Modified ZooInfoViewer to add a 
`--print-resource-groups` option, which is used in `accumulo-cluster`
to verify that a resource group is created before starting the processes.

Modified ConfigCommand to log a warning in the shell when a fixed
property is changed in a resource group.

Related to #5859


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>