Avoid two Path types in MiniAccumuloConfigImpl
Recent commits changed the way Paths/Files/URLs/URIs were converted in
MiniAccumuloConfigImpl. This changes it back to the previous way of
doing things in the main branch that avoids the confusing use of both
org.apache.hadoop.fs.Path and java.nio.file.Path in the same file, in
order to avoid rolling back the improvements done in #5547.

Instead of using both Path types, this version adds a resource as a URL
instead of as a Hadoop Path type to the Configuration.