Use ServerContext instead of Manager in fate util code
The class o.a.a.manager.tableOps.Utils took a Manager type for most of
its methods and only used the ServerContext.  Modified these methods to
take a ServerContext intead.  Making this change as the beginning of an
effort to remove direct usage of the Manager type in fate operations.