Fixes bug where monitor appender does not notice monitor died (#5972)
When the monitor process died the monitor appender did not notice
because it kept using the same stat which does not change on absence.
Also reusing the stat object may not be thread safe.  Made the stat
object a thread local to avoid object allocation and use by multiple
threads.