Handle a couple edge cases:
Handles a couple edge cases for /manager endpoint:
- If manager is dead on first loading the page (in this case, no manager
  table is created), but later comes online
  and we are using the auto-refresh feature, previously would never
create the manager table. auto-refresh will now create the table if it
does not yet exist.
- There was a case where /manager could result in a DataTables
  alert/error. If the manager was up but the endpoint used to get the
table data was not yet available (small window where this could occur),
a DataTables alert would occur. Handle this instead as a console log and
populate the table with no data.