Added summary of queued and running compactions to coordinator
This commit adds periodic logging of queued and running external
compaction information to the coordinator. The logging is emitted
by a new class, CoordinatorSummaryLogger, so that users can easily
redirect this log to a new file in the logging configuration.

At each interval this new class will log the number of compactions
running for each table, and will log the number of compactors,
queued compactions and running compactions for each compaction queue.

The number of queued compactions is an estimate as each tablet server
only reports up to 100 different compaction priorities to conserve
memory space in the Coordinator (see ExternalCompactionExecutor.summarize).

The metrics are a more accurate source of the number of queued external
compactions, but that requires aggregating all of the METRICS_MAJC_QUEUED
Meters from all of the TabletServers.

Related to #5965