Uses rendezvous hashing for scan server selection (#5937)
Modified the scan server selection plugins to use rendezvous hashing.

The ConfigurableScanServerSelector was modified to prefer servers that
have not previously failed in addition to making it use rendezvous
hashing.  This makes its behavior similar to the host based plugin.

The ConfigurableScanServerHostSelector was a hybrid between host and
server selection behavior.  For example it would select based on server
first and only on failure would it fall back to a host based selection.
Made this selector always choose based on host, even for the initial
choice. With this change the initial scans of tablets should now
randomly spread across servers on a host. Also made the host based
selector only move to the next attempt plan after all servers on all
hosts have failed.  It used to move to the next attempt plan when a
single server failed. These changes were driven by using host based
rendezvous hashing.

fixes #5928


Co-authored-by: Dave Marion <dlmarion@apache.org>