Fixed Coordinator treating compaction cancellation like failure (#6019)
The Coordinator in 2.1 checks the exception class name and if
it is not null, then treats it as a failure. In 2.1 when a
compaction is cancelled, null is passed.

In main null was removed from the Compactor and InterruptedException
was used in it's place. However the Coordinator code was not updated
and was still treating it as a failure because the exception class
name is not null. Fixed this check such that it's only treated as
a failure if the class name is not the InterruptedException class
name.

Closes #6018