avoids unneeded meta index read (#6024)
CachableBlockFile.Reader.getBCFile() always read the RFile root block
from the index cache even if would not use it. Modified the method to
only read from the index cache when the data is actually needed.
Noticed this while working on #6010, saw for every data block read from
a rfile it would also read from the index cache.

Adjusted ScanTracingIT because before this change there was an rfile
index cache read per rfile data block read.  After this change the
number of rfile index reads align w/ the number of rfiles opened, which
in this case aligns with the number of tablets in the test tables.

The existing behavior was likely not a performance problem if the data
was in cache, but could potentially cause thread contention and extra
CPU usage. So its probably good to fix, also nice to make the tracing
stats align better w/ expectations.