Reduce thread contention by shuffling rfile order (#5998)
* Shuffle the files for ServerSide iterators

Shuffle the mapFile iterators before the MultiIterator is created to
avoid block cache contention.

* Modified MultiIterator instead of ScanDataSource

Realized the changes could be moved down to MultiIterator so the
behavior was similiar for deepCopies as well as constructed
MultiIterators.

Added a test for deepCopy since one did not exist

* formatting change

* Update core/src/test/java/org/apache/accumulo/core/iterators/system/MultiIteratorTest.java

Co-authored-by: Keith Turner <kturner@apache.org>

* Adds shuffle prop and fileManager shuffling

Adds the table property for shuffling files.
Adds shuffling for files in the FileManager.

Moves the shuffling logic into a separate Iterator class and changes the
ScanDataSource code to select the specific iterator class.

* Clean up constructors for MuliIterator

Support shuffled iterators in GeneratedSplits and OfflineIterator

* fix formatting

* Remove duplicate test code

Removed the duplicate test code by extending the original test class

---------

Co-authored-by: Keith Turner <kturner@apache.org>